Make CopycatClient/CopycatServer base interfaces.
