+++ server/src/main/java/io/atomix/copycat/server/CopycatServer.java
@@ -11,7 +11,7 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
- * limitations under the License.
+ * limitations under the License
  */
 package io.atomix.copycat.server;
 
@@ -23,15 +23,10 @@ import io.atomix.catalyst.transport.Transport;
 import io.atomix.catalyst.util.Assert;
 import io.atomix.catalyst.util.ConfigurationException;
 import io.atomix.catalyst.util.Listener;
+import io.atomix.catalyst.util.Managed;
 import io.atomix.catalyst.util.concurrent.ThreadContext;
-import io.atomix.copycat.client.Command;
-import io.atomix.copycat.client.Query;
-import io.atomix.copycat.server.state.Member;
 import io.atomix.copycat.server.state.ServerContext;
-import io.atomix.copycat.server.state.ServerState;
-import io.atomix.copycat.server.storage.Log;
 import io.atomix.copycat.server.storage.Storage;
-import io.atomix.copycat.server.storage.StorageLevel;
 
 import java.time.Duration;
 import java.util.Arrays;
@@ -39,87 +34,33 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionStage;
 import java.util.function.Consumer;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 /**
- * Provides a standalone implementation of the <a href="http://raft.github.io/">Raft consensus algorithm</a>.
+ * Provides an interface for managing the lifecycle and state of a Raft server.
  * <p>
- * To create a new server, use the server {@link CopycatServer.Builder}. Servers require
- * cluster membership information in order to perform communication. Each server must be provided a local {@link Address}
- * to which to bind the internal {@link io.atomix.catalyst.transport.Server} and a set of addresses for other members in
- * the cluster.
+ * The lifecycle of the Raft server is managed via the {@link Managed} API methods. To start a server,
+ * call {@link Managed#open()} on the server. Once the server has connected to the cluster and found
+ * a leader, the returned {@link CompletableFuture} will be completed and the server will be operating.
  * <p>
- * Underlying each server is a {@link StateMachine}. The state machine is responsible for maintaining the state with
- * relation to {@link Command}s and {@link Query}s submitted to the
- * server by a client.
- * <pre>
- *   {@code
- *   Address address = new Address("123.456.789.0", 5000);
- *   Collection<Address> members = Arrays.asList(new Address("123.456.789.1", 5000), new Address("123.456.789.2", 5000));
+ * Throughout the lifetime of a server, the server transitions between a variety of
+ * {@link CopycatServer.State states}. Call {@link #state()} to get the current state
+ * of the server at any given point in time.
+ * <p>
+ * The {@link #term()} and {@link #leader()} are critical aspects of the Raft consensus algorithm. Initially,
+ * when the server is started, the {@link #term()} will be initialized to {@code 0} and {@link #leader()} will
+ * be {@code null}. By the time the server is fully started, both {@code term} and {@code leader} will be
+ * provided. As the cluster progresses, {@link #term()} will progress monotonically, and for each term,
+ * only a single {@link #leader()} will ever be elected.
+ * <p>
+ * Raft servers are members of a cluster of servers identified by their {@link Address}. Each server
+ * must be able to locate other members of the cluster. Throughout the lifetime of a cluster, the membership
+ * may change. The {@link #members()} method provides a current view of the cluster from the perspective
+ * of a single server. Note that the members list may not be consistent on all nodes at any given time.
  *
- *   CopycatServer server = CopycatServer.builder(address, members)
- *     .withStateMachine(new MyStateMachine())
- *     .build();
- *   }
- * </pre>
- * Server state machines are responsible for registering {@link Command}s which can be submitted
- * to the cluster. Raft relies upon determinism to ensure consistency throughout the cluster, so <em>it is imperative
- * that each server in a cluster have the same state machine with the same commands.</em>
- * <p>
- * By default, the server will use the {@code NettyTransport} for communication. You can configure the transport via
- * {@link CopycatServer.Builder#withTransport(Transport)}.
- * <p>
- * As {@link Command}s are received by the server, they're written to the Raft {@link Log}
- * and replicated to other members of the cluster. By default, the log is stored on disk, but users can override the default
- * {@link Storage} configuration via {@link CopycatServer.Builder#withStorage(Storage)}. Most notably,
- * to configure the storage module to store entries in memory instead of disk, configure the
- * {@link StorageLevel}.
- * <pre>
- * {@code
- * CopycatServer server = CopycatServer.builder(address, members)
- *   .withStateMachine(new MyStateMachine())
- *   .withStorage(new Storage(StorageLevel.MEMORY))
- *   .build();
- * }
- * </pre>
- * All serialization is performed with a Catalyst {@link Serializer}. By default, the serializer loads registered
- * {@link io.atomix.catalyst.serializer.CatalystSerializable} types with {@link ServiceLoaderTypeResolver}, but users
- * can provide a custom serializer via {@link CopycatServer.Builder#withSerializer(Serializer)}.
- * The server will still ensure that internal serializable types are properly registered on user-provided serializers.
- * <p>
- * Once the server has been created, to connect to a cluster simply {@link #open()} the server. The server API is
- * fully asynchronous and relies on {@link CompletableFuture} to provide promises:
- * <pre>
- * {@code
- * server.open().thenRun(() -> {
- *   System.out.println("Server started successfully!");
- * });
- * }
- * </pre>
- * When the server is started, it will attempt to connect to an existing cluster. If the server cannot find any
- * existing members, it will attempt to form its own cluster.
- * <p>
- * Once the server is started, it will communicate with the rest of the nodes in the cluster, periodically
- * transitioning between states. Users can listen for state transitions via {@link #onStateChange(Consumer)}:
- * <pre>
- * {@code
- * server.onStateChange(state -> {
- *   if (state == CopycatServer.State.LEADER) {
- *     System.out.println("Server elected leader!");
- *   }
- * });
- * }
- * </pre>
- *
- * @see StateMachine
- * @see Transport
- * @see Storage
- * @author <a href="http://github.com/kuujo">Jordan Halterman</a>
+ * @author <a href="http://github.com/kuujo>Jordan Halterman</a>
  */
-public class CopycatServer implements RaftServer {
+public interface CopycatServer extends Managed<CopycatServer> {
 
   /**
    * Returns a new Raft server builder.
@@ -138,7 +79,7 @@ public class CopycatServer implements RaftServer {
    * @param cluster The cluster members to which to connect.
    * @return The server builder.
    */
-  public static Builder builder(Address clientAddress, Address serverAddress, Address... cluster) {
+  static Builder builder(Address clientAddress, Address serverAddress, Address... cluster) {
     return builder(clientAddress, serverAddress, Arrays.asList(cluster));
   }
 
@@ -159,36 +100,86 @@ public class CopycatServer implements RaftServer {
    * @param cluster The cluster members to which to connect.
    * @return The server builder.
    */
-  public static Builder builder(Address clientAddress, Address serverAddress, Collection<Address> cluster) {
+  static Builder builder(Address clientAddress, Address serverAddress, Collection<Address> cluster) {
     return new Builder(clientAddress, serverAddress, cluster);
   }
 
-  private final ServerContext context;
-  private CompletableFuture<RaftServer> openFuture;
-  private CompletableFuture<Void> closeFuture;
-  private ServerState state;
-  private final Duration electionTimeout;
-  private final Duration heartbeatInterval;
-  private final Duration sessionTimeout;
-  private Listener<Address> electionListener;
-  private boolean open;
+  /**
+   * Raft server state types.
+   * <p>
+   * States represent the context of the server's internal state machine. Throughout the lifetime of a server,
+   * the server will periodically transition between states based on requests, responses, and timeouts.
+   *
+   * @author <a href="http://github.com/kuujo">Jordan Halterman</a>
+   */
+  enum State {
 
-  private CopycatServer(ServerContext context, Duration electionTimeout, Duration heartbeatInterval, Duration sessionTimeout) {
-    this.context = context;
-    this.electionTimeout = electionTimeout;
-    this.heartbeatInterval = heartbeatInterval;
-    this.sessionTimeout = sessionTimeout;
+    /**
+     * Represents the state of an inactive server.
+     * <p>
+     * All servers start in this state and return to this state when {@link #close() stopped}.
+     */
+    INACTIVE,
+
+    /**
+     * Represents the state of a server that does not store state.
+     */
+    RESERVE,
+
+    /**
+     * Represents the state of a server in the process of catching up its log.
+     * <p>
+     * Upon successfully joining an existing cluster, the server will transition to the passive state and remain there
+     * until the leader determines that the server has caught up enough to be promoted to a full member.
+     */
+    PASSIVE,
+
+    /**
+     * Represents the state of a server participating in normal log replication.
+     * <p>
+     * The follower state is a standard Raft state in which the server receives replicated log entries from the leader.
+     */
+    FOLLOWER,
+
+    /**
+     * Represents the state of a server attempting to become the leader.
+     * <p>
+     * When a server in the follower state fails to receive communication from a valid leader for some time period,
+     * the follower will transition to the candidate state. During this period, the candidate requests votes from
+     * each of the other servers in the cluster. If the candidate wins the election by receiving votes from a majority
+     * of the cluster, it will transition to the leader state.
+     */
+    CANDIDATE,
+
+    /**
+     * Represents the state of a server which is actively coordinating and replicating logs with other servers.
+     * <p>
+     * Leaders are responsible for handling and replicating writes from clients. Note that more than one leader can
+     * exist at any given time, but Raft guarantees that no two leaders will exist for the same {@link #term()}.
+     */
+    LEADER
+
   }
 
-  @Override
-  public long term() {
-    return state.getTerm();
-  }
+  /**
+   * Returns the current Raft term.
+   * <p>
+   * The term is a monotonically increasing number that essentially acts as a logical time for the cluster. For any
+   * given term, Raft guarantees that only one {@link #leader()} can be elected, but note that a leader may also
+   * not yet exist for the term.
+   *
+   * @return The current Raft term.
+   */
+  long term();
 
-  @Override
-  public Address leader() {
-    return state.getLeader().serverAddress();
-  }
+  /**
+   * Returns the current Raft leader.
+   * <p>
+   * If no leader has been elected, the leader address will be {@code null}.
+   *
+   * @return The current Raft leader or {@code null} if this server does not know of any leader.
+   */
+  Address leader();
 
   /**
    * Registers a leader election listener.
@@ -203,19 +194,29 @@ public class CopycatServer implements RaftServer {
    * {@link Listener#close()}.
    * @throws NullPointerException If {@code listener} is {@code null}
    */
-  public Listener<Address> onLeaderElection(Consumer<Address> listener) {
-    return state.onLeaderElection(listener);
-  }
+  Listener<Address> onLeaderElection(Consumer<Address> listener);
 
-  @Override
-  public Collection<Address> members() {
-    return state.getMembers().stream().map(Member::serverAddress).collect(Collectors.toList());
-  }
+  /**
+   * Returns a collection of current cluster members.
+   * <p>
+   * The current members list includes members in all states, including non-voting states. Additionally, because
+   * the membership set can change over time, the set of members on one server may not exactly reflect the
+   * set of members on another server at any given point in time.
+   *
+   * @return A collection of current Raft cluster members.
+   */
+  Collection<Address> members();
 
-  @Override
-  public State state() {
-    return state.getState();
-  }
+  /**
+   * Returns the Raft server state.
+   * <p>
+   * The initial state of a Raft server is {@link State#INACTIVE}. Once the server is {@link #open() started} and
+   * until it is explicitly shutdown, the server will be in one of the active states - {@link State#PASSIVE},
+   * {@link State#FOLLOWER}, {@link State#CANDIDATE}, or {@link State#LEADER}.
+   *
+   * @return The Raft server state.
+   */
+  State state();
 
   /**
    * Registers a state change listener.
@@ -238,110 +239,65 @@ public class CopycatServer implements RaftServer {
    * {@link Listener#close()}.
    * @throws NullPointerException If {@code listener} is {@code null}
    */
-  public Listener<State> onStateChange(Consumer<State> listener) {
-    return state.onStateChange(listener);
-  }
-
-  @Override
-  public ThreadContext context() {
-    return state.getThreadContext();
-  }
-
-  @Override
-  public CompletableFuture<RaftServer> open() {
-    if (open)
-      return CompletableFuture.completedFuture(this);
-
-    if (openFuture == null) {
-      synchronized (this) {
-        if (openFuture == null) {
-          Function<ServerState, CompletionStage<RaftServer>> completionFunction = state -> {
-            CompletableFuture<RaftServer> future = new CompletableFuture<>();
-            openFuture = null;
-            this.state = state;
-            state.setElectionTimeout(electionTimeout)
-              .setHeartbeatInterval(heartbeatInterval)
-              .setSessionTimeout(sessionTimeout)
-              .join()
-              .whenComplete((result, error) -> {
-                if (error == null) {
-                  if (state.getLeader() != null) {
-                    open = true;
-                    future.complete(this);
-                  } else {
-                    electionListener = state.onLeaderElection(leader -> {
-                      if (electionListener != null) {
-                        open = true;
-                        future.complete(this);
-                        electionListener.close();
-                        electionListener = null;
-                      }
-                    });
-                  }
-                } else {
-                  future.completeExceptionally(error);
-                }
-              });
-            return future;
-          };
-
-          if (closeFuture == null) {
-            openFuture = context.open().thenCompose(completionFunction);
-          } else {
-            openFuture = closeFuture.thenCompose(c -> context.open().thenCompose(completionFunction));
-          }
-        }
-      }
-    }
-    return openFuture;
-  }
-
-  @Override
-  public boolean isOpen() {
-    return open;
-  }
-
-  @Override
-  public CompletableFuture<Void> close() {
-    if (!open)
-      return CompletableFuture.completedFuture(null);
-
-    if (closeFuture == null) {
-      synchronized (this) {
-        if (closeFuture == null) {
-          if (openFuture == null) {
-            closeFuture = state.leave()
-              .thenCompose(v -> context.close())
-              .whenComplete((result, error) -> state = null);
-          } else {
-            closeFuture = openFuture.thenCompose(c -> state.leave()
-              .thenCompose(v -> context.close()))
-              .whenComplete((result, error) -> state = null);
-          }
-        }
-      }
-    }
-    return closeFuture;
-  }
-
-  @Override
-  public boolean isClosed() {
-    return !open;
-  }
+  Listener<State> onStateChange(Consumer<State> listener);
 
   /**
-   * Deletes the Raft server and its logs.
+   * Returns the server execution context.
+   * <p>
+   * The thread context is the event loop that this server uses to communicate other Raft servers.
+   * Implementations must guarantee that all asynchronous {@link java.util.concurrent.CompletableFuture} callbacks are
+   * executed on a single thread via the returned {@link io.atomix.catalyst.util.concurrent.ThreadContext}.
+   * <p>
+   * The {@link io.atomix.catalyst.util.concurrent.ThreadContext} can also be used to access the Raft server's internal
+   * {@link io.atomix.catalyst.serializer.Serializer serializer} via {@link ThreadContext#serializer()}. Catalyst serializers
+   * are not thread safe, so to use the context serializer, users should clone it:
+   * <pre>
+   *   {@code
+   *   Serializer serializer = server.threadContext().serializer().clone();
+   *   Buffer buffer = serializer.writeObject(myObject).flip();
+   *   }
+   * </pre>
+   *
+   * @return The server thread context.
+   */
+  ThreadContext context();
+
+  /**
+   * Starts the Raft server asynchronously.
+   * <p>
+   * When the server is started, the server will attempt to search for an existing cluster by contacting all of
+   * the members in the provided members list. If no existing cluster is found, the server will immediately transition
+   * to the {@link State#FOLLOWER} state and continue normal Raft protocol operations. If a cluster is found, the server
+   * will attempt to join the cluster. Once the server has joined or started a cluster and a leader has been found,
+   * the returned {@link CompletableFuture} will be completed.
+   *
+   * @return A completable future to be completed once the server has joined the cluster and a leader has been found.
+   */
+  @Override
+  CompletableFuture<CopycatServer> open();
+
+  /**
+   * Returns a boolean value indicating whether the server is running.
+   * <p>
+   * Once {@link #open()} is called and the returned {@link CompletableFuture} is completed (meaning this server found
+   * a cluster leader), this method will return {@code true} until {@link #close() closed}.
+   *
+   * @return Indicates whether the server is running.
+   */
+  @Override
+  boolean isOpen();
+
+  /**
+   * Deletes the Copycat server and its logs.
    *
    * @return A completable future to be completed once the server has been deleted.
    */
-  public CompletableFuture<Void> delete() {
-    return close().thenRun(context::delete);
-  }
+  CompletableFuture<Void> delete();
 
   /**
    * Raft server builder.
    */
-  public static class Builder extends io.atomix.catalyst.util.Builder<CopycatServer> {
+  class Builder extends io.atomix.catalyst.util.Builder<CopycatServer> {
     private static final Duration DEFAULT_RAFT_ELECTION_TIMEOUT = Duration.ofMillis(1000);
     private static final Duration DEFAULT_RAFT_HEARTBEAT_INTERVAL = Duration.ofMillis(150);
     private static final Duration DEFAULT_RAFT_SESSION_TIMEOUT = Duration.ofMillis(5000);
@@ -520,7 +476,7 @@ public class CopycatServer implements RaftServer {
       }
 
       ServerContext context = new ServerContext(clientAddress, serverAddress, cluster, quorumHint, backupCount, stateMachine, transport, storage, serializer);
-      return new CopycatServer(context, electionTimeout, heartbeatInterval, sessionTimeout);
+      return new CopycatRaftServer(context, electionTimeout, heartbeatInterval, sessionTimeout);
     }
   }
