+++ server/src/main/java/io/atomix/copycat/server/storage/entry/OperationEntry.java
@@ -15,11 +15,11 @@
  */
 package io.atomix.copycat.server.storage.entry;
 
-import io.atomix.copycat.client.Operation;
 import io.atomix.catalyst.buffer.BufferInput;
 import io.atomix.catalyst.buffer.BufferOutput;
 import io.atomix.catalyst.serializer.Serializer;
 import io.atomix.catalyst.util.ReferenceManager;
+import io.atomix.copycat.client.Operation;
 
 /**
  * Operation entry.
