+++ server/src/main/java/io/atomix/copycat/server/storage/entry/QueryEntry.java
@@ -15,14 +15,14 @@
  */
 package io.atomix.copycat.server.storage.entry;
 
-import io.atomix.copycat.client.Operation;
-import io.atomix.copycat.client.Query;
 import io.atomix.catalyst.buffer.BufferInput;
 import io.atomix.catalyst.buffer.BufferOutput;
 import io.atomix.catalyst.serializer.SerializeWith;
 import io.atomix.catalyst.serializer.Serializer;
 import io.atomix.catalyst.util.Assert;
 import io.atomix.catalyst.util.ReferenceManager;
+import io.atomix.copycat.client.Operation;
+import io.atomix.copycat.client.Query;
 
 /**
  * Query entry.
