+++ server/src/test/java/io/atomix/copycat/server/state/ActiveStateTest.java
@@ -16,7 +16,7 @@
 package io.atomix.copycat.server.state;
 
 import io.atomix.copycat.client.response.Response.Status;
-import io.atomix.copycat.server.RaftServer.State;
+import io.atomix.copycat.server.CopycatServer.State;
 import io.atomix.copycat.server.request.AppendRequest;
 import io.atomix.copycat.server.request.PollRequest;
 import io.atomix.copycat.server.request.VoteRequest;
