+++ server/src/test/java/io/atomix/copycat/server/state/CandidateStateTest.java
@@ -17,7 +17,7 @@ package io.atomix.copycat.server.state;
 
 import io.atomix.catalyst.transport.Server;
 import io.atomix.copycat.client.response.Response;
-import io.atomix.copycat.server.RaftServer;
+import io.atomix.copycat.server.CopycatServer;
 import io.atomix.copycat.server.request.AppendRequest;
 import io.atomix.copycat.server.request.VoteRequest;
 import io.atomix.copycat.server.response.AppendResponse;
@@ -59,7 +59,7 @@ public class CandidateStateTest extends AbstractStateTest<CandidateState> {
       assertEquals(serverState.getTerm(), 2L);
       assertEquals(serverState.getLeader().id(), leader);
       assertEquals(response.term(), 2L);
-      assertEquals(serverState.getState(), RaftServer.State.FOLLOWER);
+      assertEquals(serverState.getState(), CopycatServer.State.FOLLOWER);
     });
   }
 
@@ -124,7 +124,7 @@ public class CandidateStateTest extends AbstractStateTest<CandidateState> {
       assertEquals(serverState.getTerm(), 3L);
       assertEquals(serverState.getLastVotedFor(), candidate);
       assertEquals(response.term(), 3L);
-      assertEquals(serverState.getState(), RaftServer.State.FOLLOWER);
+      assertEquals(serverState.getState(), CopycatServer.State.FOLLOWER);
     });
   }
 
@@ -153,13 +153,13 @@ public class CandidateStateTest extends AbstractStateTest<CandidateState> {
       assertEquals(serverState.getTerm(), 3L);
       assertEquals(serverState.getLastVotedFor(), 0);
       assertEquals(response.term(), 3L);
-      assertEquals(serverState.getState(), RaftServer.State.FOLLOWER);
+      assertEquals(serverState.getState(), CopycatServer.State.FOLLOWER);
     });
   }
 
   public void testCandidateTransitionsToLeaderOnElection() throws Throwable {
     serverState.onStateChange(state -> {
-      if (state == RaftServer.State.LEADER)
+      if (state == CopycatServer.State.LEADER)
         resume();
     });
 
@@ -191,7 +191,7 @@ public class CandidateStateTest extends AbstractStateTest<CandidateState> {
 
   public void testCandidateTransitionsToFollowerOnRejection() throws Throwable {
     serverState.onStateChange(state -> {
-      if (state == RaftServer.State.FOLLOWER)
+      if (state == CopycatServer.State.FOLLOWER)
         resume();
     });
