+++ server/src/test/java/io/atomix/copycat/server/storage/SegmentManagerTest.java
@@ -1,7 +1,5 @@
 package io.atomix.copycat.server.storage;
 
-import static org.testng.Assert.assertEquals;
-
 import org.testng.annotations.Test;
 
 /**
